filter out dom props
