+++ src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './tailSafe.utils';
+export * from './dom.utils';
